Add prodgt0 to iset.mm
Includes prodgt0gt0 which is a lemma in the sense that prodgt0
is stronger, but which doesn't need to be named with "lem" in the
name because it isn't too hard to see a proof saving a few steps by
using prodgt0gt0 rather than prodgt0 .  The proof is taken fairly
directly from part of the prodgt0 proof from set.mm.

prodgt0 is stated as in set.mm. The proof is not particularly
similar to set.mm's proof, though.