Encourage simpl and simpr names in iset.mm
When we have a statement which is a theorem in set.mm and an
axiom in iset.mm our general practice is to include the set.mm
name as a theorem and mark the axiom as discouraged, so that
the axiom name only gets used in proving that theorem.

Apply this with ax-ia1 (also known as simpl ) and ax-ia2 (also
known as simpr ).

Rename existing uses and rewrap.