Intuitionize qreccl in iset.mm
The proof is similar to the set.mm one except for a large amount
of converting between not equal and apart. Since not equal and
apart are equivalent for rationals - see qapne - we keep the
statement of qreccl as in set.mm rather than stating it in terms
of apartness.