Add rpnegap to iset.mm
It is similar to rpneg with the statement of the theorem (a) changing
not equal to apart, and (b) stating exclusive or using \/_ rather
than via an expression which is equivalent given excluded middle.

The proof is similar to the set.mm one in the sense that it relies
on elrp (twice) and lt0neg1 , although the logic is fairly different.