Add ixxss1 and iooss1 to iset.mm
These replace ixxss1g and iooss1g and bring us closer to set.mm.

We keep the $d x y z O $. for ixxssxr (which doesn't match set.mm
but we'll keep it unless we need to do the variable change to
remove it) but omit it for elixx3g (where some of the theorems
which rely on it don't have the distinct variable constraint).

Copy the ioodisj proof from set.mm without changes.