Modify mm_100.html to add etransc
Modify the "Metamath 100" file mm_100.html to record the new
proof etransc, e is Transcendental, proof 67 in the Metamath 100.

A big *congrats* to Glauco Siliprandi for completing this proof!

This commit updates the Metamath 100 total to 74 (instead of 73).
I updated the other prover totals as well per Freek's list.
Since I'm editing file mm_100.html anyway, I also specifically
listed the five different verifiers that are used on set.mm.
I think this is a distinctive of Metamath compared to many other
systems, so it's worth providing a little more information about it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>