Add indstr to iset.mm
The proof in set.mm depends on various non-intuitionistic pieces
of logic. So instead we just go for the straightforward proof of
strong induction where we apply regular induction on the statement
that the induction hypothesis holds up to a given value. There
is nothing especially tricky about this, although formalizing it (at
least the way we do it) is a bit on the verbose side.

Also includes zleloe which is an easy consequence of integer
trichotomy.