Add nn01to3 to iset.mm
The theorem is as stated in set.mm. The proof is based on zleloe
and nn0ltp1le . This seemed like it might be a bit easier than
trying to intuitionize each of the steps in the set.mm proof,
(a number of them would need changes). The result is 855
characters compared with 987 characters for the set.mm proof
so I guess it is a success.