Add some stoic logic theorems to iset.mm
This is stoic2a , stoic2b , stoic3 , stoic4a , and stoic4b .
These are copied without change from set.mm.

Although this doesn't fully update this section to the current
set.mm, it does get us closer and fills in theorems at
least some of which we'll want to use in iset.mm.