Extend comments on Tarski geometry axioms
Extend some comments to help explain some of the Tarski geometry axioms.

I felt that some of the current Tarski geometry axioms in this database
were not easy to understand without going back to the original source
materials. This commit adds comments to help people (including me!)
understand these axioms if they don't go back to the source materials
(much of which are in German).

This commit also adds a citation to a 1999 paper published in The Bulletin
of Symbolic Logic.  That 1999 paper is in English, which is easier to
understand for those who can read English but not German. That paper also
provides additional background and context, as well as some
intuitive explanations which are quoted (with credit). In cases the symbols
in the paper are different from this database, the quotes were modified
(for clarity) so that they match the symbols used in this database.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>