Add aacllem to David A. Wheeler's mathbox
Restore theorem aacllem by Brendan Leahy, with fixes by
Alexander van der Vekens and David A. Wheeler to make it work.
David A. Wheeler also minimized it (rspa shortens it slightly).

This is a complex proof that I didn't want to see lost.
I put this in my mathbox as a start, if Brendan Leahy wants it
moved to his mathbox, of course feel free to move it!

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>