Change Gource caption input dates to ISO format
Change all Gource caption date input to ISO format YYYY-MM-DD.

Gource wants time in Unix timestamp format
(number of seconds since 1970-01-01 ignoring leap seconds).
However, that's a pain to use, easy to make a mistake in,
and hard to review.

For example, the timestamp for the book was slightly wrong
(claimed 1559448000, should have been 1559433600).
That didn't matter in the grand scheme of things, but it
was an indicator that something needed to change.

This commit changes the Gource caption input to use ISO date
format YYYY-MM-DD, which is easy to work with programmatically
(for example, it sorts easily) AND it's easy to review.
We later convert dates to Unix timestamps just before sending the
data to Gource.

This makes editing and reviewing all date-related caption
information much easier and less prone to error.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>