Scripts now download and install Metamath from GitHub
Modify file scripts/download-metamath so that it downloads the
currently-released version of the Metamath source from GitHub,
and modify scripts/build-metamath so that it can build from
that version.

The "latest" version is the git tag on GitHub that is the
last in a natural sort (sequences of numbers are compared as numbers).
For example, v10.2 > v9.3. Note that you can add a missing older tag
without harm, because we sort the version numbers not the date the
version numbers were added.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>