Move symmetric difference into main body
Move the symmetric difference operation on sets
into the main body, along with a few proofs about it.

Symmetric difference is a standard operation on sets.
There's even a Wikipedia article specifically about it
(<https://en.wikipedia.org/wiki/Symmetric_difference>).
There was a long discussion on the Metamath mailing list
on how to represent it typographically, which also suggests
that people care about this operation. Our usual rule moving
things is when more than one mathbox uses it, and now we have
multiple users.  Thus, it makes sense to move it to the main body.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>