Reserve a set of constants for ASCII characters
Reserve a set of constants for ASCII characters
in David A. Wheeler's mathbox.

We may someday want to represent ASCII characters as constants
(e.g., for "Metamath in Metamath"). Mario is already doing
this in an MM0 database.

This commit creates a proposed set of constant names
to "reserve" these names for characters in set.mm.
That way, if we choose to use them, this set of symbols
will be reserved and ready for their use.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>