scripts/gource-captions.txt: prefix "type" labels for uniformity
In gource-captions.txt, add a "type" label prefix
to indicate the type of event. Also add the date to the
*visible* section (do this by just repeating the date).

This is per a suggestion by Benoit Jubin on the Metamath mailing list,
1 May 2020 05:09:07 -0700 (PDT), subject
"Re: [Metamath] Last-minute comments on updated Gource visualization?".

I also add a reference to the first Gource visualization, which
is absurdly self-referential, but why not :-).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>