Add frec2uz0i to iset.mm
Similar to om2uz0i in set.mm but stated in terms of frec instead
of rec restricted to _om . Also changes the domain of the
characteristic function from _V to ZZ .

This is the first theorem in the section "Miscellaneous theorems
about integers" from set.mm. The rest of this section will need
additional work (perhaps somewhat similar in spirit).