Modify finite recursion in iset.mm
The change here is to change F Fn _V to A. y ( F ` y ) e. _V
throughout this section.

Affected theorems are frecabex , frectfr , frecfnom , frecsuclem1 ,
frecsuclemdm , frecsuclem2 , frecsuclem3 , and frecsuc .

Also adds rdgtfr and rdgifnon2 which are similar for rec rather than
frec . However, we'd need to change over more rec-related theorems
for frecrdg to be able to get rid of F Fn _V , so we keep F Fn _V
in the statement of frecrdg and adjust the proof of frecrdg to
adapt to the changes to frecsuc .

Reword the comment for rdgifnon ; there is no requirement that the
result of the characteristic function has to be an ordinal.