Add ssfiexmid to iset.mm
It fits in with the finite set theorems I have been adding lately,
and turns out not to be too hard. The proof is similar to the
one in Bauer although the details are slightly different as it
doesn't enumerate the set as such.

Also add enm which I needed for this proof.