Rename frecuzrdgfn etc to *OLD in iset.mm
This is most of the theorems related to seq . The future change
will be to change ZZ to ( ZZ>= ` M ) in the closure law
  ( ph /\ x e. ZZ ) -> ( F ` x ) e. S
(where M is the initial index), which appears in all of these
theorems in one form or another. Renaming the ones which use
ZZ to *OLD is the first step.