Add exponentiation to iset.mm
This is df-exp and expival .

df-exp is only adjusted for the different syntax of seq .

expival is similar to expval but adds a condition to exclude the
case in which the base is zero and the exponent is negative.

There's also a lemma expivallem which helps with the case where
the base is apart from zero.