Add expnbnd to iset.mm
Although it may not be easy to see it, the proof is quite similar
to the set.mm one. Taking the floor and adding one is changed to
arch , and ltlecasei is changed to axltwlin . But although those
changes affect the way a lot of steps look, they don't change the
basic division into two cases, using an exponent of one for the
A < B case, and using bernneq2 for the 1 < A case.