Remove duplicate logblogfv and improve elogb
The proofs logblogfv and elogb duplicate each other; we only need one.

The elogb proof was first (contributed in 2017) had shorter proofs
for the cases e =/= 1 and e =/= 0 and made those proofs reusable.
However, logblogfv had a shorter proof for building on those facts than
the original elogb proof.

This commit removes logblogfv and modifies elogb so it has the
best traits of the two original proofs (it's one step shorter now).
It also combines the credits (since it's now a combined work of DAW and AV).

This commit also rewraps set.mm using "scripts/rewrap".

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>