Use two variables in nqprlu
Rename nqprlu to nqprxx and add nqprlu which has a variable
change to use different variables for the lower cut and upper cut.
Rename usages of nqprlu to nqprxx (and switch it back in
some cases - by running the minimizer on all theorems which had
used the old nqprlu).

Although it would be possible to bundle the two theorems by
removing the distinct variable constraint between l and u in
nqprlu , I don't know whether that is considered preferable or
not.