Add actual Schwabhauser 2.5
Although tgcgrcomlr is true (congruence commutes on both sides),
that is not Schwabhauser theorem 2.5 (congruence commutes on the RHS).

This commit proves Schwabhauser theorem 2.5 (implication that
congruence commutes on the RHS) as it is stated in the book.
This commit also adds two simple related theorems that commute
congruence on the RHS and LHS; congruence is an extremely important
relationship in geometry, so it's convenient to be able
to commute congruence in one step. This commit also fixes
the commment of tgcgrcomlr so that it no longer says it is theorem 2.5.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>