Expand df-fr discussion in mmil.html
On the whole, it seems better to describe them here (as they
certainly do represent a difference between set.mm and iset.mm).
We also get better typesetting and linking to theorems than we
would in a github issue, and there is a small amount of updating
based on what has been proved since github issue #646 was written.