Prove caucvgprlemloc in iset.mm
The proof is similar to cauappcvgprlemloc with x changed to 1 / n
in most parts of the proof. We also avoid distinct variable
constraints between L and the variables on the right hand side
of caucvgpr.lim which will mean that theorems which use this
lemma won't need a variable change there.