New approach for graph theory
See also discussion in issue #1418:
Since the beginning of dealing with graph theory in set.mm, the question if graphs should be represented as ordered pairs (of a set of vertices and a set of edges) or as extensible structure (the set of vertices as base set, and a special slot for the edges) is open (although the current graph theory is based on the representation by ordered pairs).
By this PR, I provid definitions ~df-vtx and ~df-iedg for functions `Vtx` and `iEdg` to get the vertices and (indexed) edges of graphs represented by an arbitrary class G, and this works for G being an ordered pair or an extensible structures (and some other, maybe degenerated cases).  If required, other representations could be supported by extending the definitions in the future.
The goal is now to build up the graph theory on these two functions, disregarding the concrete representation of a graph. Therefore, a decision how to represent a graph will be not necessary anymore.