Check pull requests without duplicate checks
Also check pull requests (as well as pushes).
This is necessary when people create a pull request for a change
they made within their own repo.

The "easy solution" in GitHub Actions is to do this:

~~~~
on: [push, pull_request]
~~~~

However, this causes duplicates in some circumstances. E.g.,
"[If we get a pull request from a forked repo, the CI runs correctly.
The problem is, if we create a pull request ourselves,
we get duplicated  checks on all items"
Source: [dups](https://github.community/t/duplicate-checks-on-push-and-pull-request-simultaneous-event/18012/7)

We resolve this by using "skip-duplicate-actions" action from:
https://github.com/marketplace/actions/skip-duplicate-actions

Note - this also adds a "workflow_dispatch" event, so that these
jobs can also be easily triggered at any time directly from the
GitHub API.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>