Remove iset.mm axiom numbering comments
Some of the axioms had comments like "Axiom 1 of 10 for
intuitionistic logic" but (a) these have not been kept up to date
as axioms have been added and removed, and (b) even in the original
concept, they didn't try to include axioms which were also in
set.mm, either as-is or in somewhat modified fashion.

I suppose if we want to come up with some numbering system later
we can do that (although I'm not sure whether the axiom set has
reached its "final" form yet), but there's not much harm in
getting rid of some numbers which as far as I know no one has
been paying much attention to.