Fix algcvgb and its lemma in iset.mm
Rename ialgcvgb back to algcvgb . Neither the statement of the
theorem nor the proof needs to be any different in iset.mm as
compared with set.mm.

Rename ialgcvgblem back to algcvgblem .  The proof is different
from set.mm but the statement of the theorem is the same as
set.mm.