Add climshft to iset.mm
The statement of climshft is as in set.mm. The proof is intutionized
by changing ovex to ovshftex and climshftlem to climshftlemg.

climshftlemg is similar to climshftlem in set.mm except that the A e. _V
condition is an antecedent instead of a hypothesis.

climres is copied from set.mm without changes.