Add iseqex to iset.mm
The statement of the theorem is the same as seqex in set.mm except
that seq has an additional argument in iset.mm. The proof is
adjusted for the differing definition seq but turns out not to be hard.

Also add tfrfun , a theorem which says that recs produces a function.

Also add frecex , a theorem which says that frec produces a set.