Add sumeq2d to iset.mm
This is based on sumeq2ii in set.mm but is proved in a somewhat
different way because of the modified definition of sum_ (which is also
modified in this change). It drops the part about guarding with _I
(because that is somewhat lightly used in set.mm and didn't go as
smoothly here as in set.mm although some of it could perhaps be
added if we have a need in the future).

Includes sumdc which is a way of restating the decidability condition in
sumeq2d .

Changes the definition of sum_ to add (a) a decidability condition to
the series case, and (b) an if statement to the finite case to ensure we
are evaluating functions within their domains. Rename from df-sum to
df-isum because I guess that's what we do when we change the definition.

Change the proofs of sumeq1 , nfsum , and nfsum1 for the new definition
of sum_ .