Remove onminex comment in mmil.html
This is evidently some flavor of copy-paste from oneqmin (for one
thing, onminex isn't a biconditional, so "reverse direction" doesn't
make sense here). I went back to the github issues and the only
hit in a search was for https://github.com/metamath/set.mm/issues/629
which doesn't have much discussion or detail, so seems better to
just not say anything for now.