Add phpm to iset.mm
This is similar to php in set.mm.  The statement of the theorem is
changed to change what we mean by "proper subset". The proof is
relatively similar but rearranged a lot and uses php5dom in place of a
number of steps which set.mm performs with sbth and other theorems.