Add df-fl and df-ceil to iset.mm
Move a bit of the text about how this can be tricky from mmil.html to
the df-fl comment (we'll see as we go exactly how many things that
affects - in some cases it may be possible to show that a floor exists).

Also add flval which is stated as in set.mm and has a similar proof
(mainly need to change riotaex to riotaexg ).