Add comments regarding convergence in iset.mm
Add comments to cauappcvgpr , caucvgpr , and caucvgprpr which
mention the other two. I'm not sure how strongly to suggest a particular
study order (or even whether to suggest studying this proof since it is
so long and so buried in inconvenient notation), but I wanted to say
something about why we have three versions of basically the same
proof (when only one is being used), and this was the best I could
think of.