Add addmodlteq to iset.mm
The theorem is stated as in set.mm. The proof changes real to rational
or integer in various places and also rearranges the logic in an effort
to simplify it.

The set.mm proof is 4131 characters in show proof/size. The iset.mm
proof is 3627 characters, so the rearrangement was apparently
successful.