Intuitionize dvdsle in iset.mm
The statement of dvdsle does not change. The proof is rearranged to use
integers or rationals rather than reals.

The lemma dvdslelem is renamed to dvdslelemd and recast into deduction
form, which means that the dvdsle proof does not need the deduction
theorem. The dvdslelemd proof also is modified to use rationals rather
than reals in places.