Add zeo3 to iset.mm
The theorem is stated as in set.mm. The proof is different from set.mm's
but isn't hard from zeo and zeo2 (a shorter proof might be
possible, but this one isn't super long).

Also add zeo4 which is stated as in set.mm but proved from zeo3 .
Add a new theorem zeoxor which is one of the steps to proving
zeo4 but seems like it may be worth stating as (arguably) a more
faithful expression of "either odd or even but not both" than zeo4 (in
the absence of excluded middle).