Add oddpwdc to iset.mm
This is stated as in set.mm. The proof is from pw2dvds instead of the
set.mm proof.

Includes lemma pw2dvdseu which is pw2dvds plus uniqueness.

Includes lemma pw2dvdseulemle which is a theorem about divisibility by
powers of two, which is used in the proof of pw2dvdseu .

Includes lemma oddpwdclemdc which is the main part of the oddpwdc
proof.

Includes oddpwdclemxy which is much of the oddpwdclemdc proof.

Includes oddpwdclemodd which states that removing the powers of
two produces an odd number.

Includes lemma oddpwdclemdvds which expands the iota to say its
result divides the natural number which generates the iota. Includes
lemma oddpwdclemndvds which is the same thing for the other half of the
iota (that is, the next power of two does not divide it).