Add divalg to iset.mm
The theorem is stated as in set.mm.

The proof is very different because the set.mm one is in terms of
infimum and does not seem like it could easily be salvaged.

Includes lemmas divalglemnn , divalglemeunn , divalglemqt ,
divalglemnqt , divalglemeuneg , and divalglemex which are for special
cases or parts of the proof.