Add supclti to iset.mm
This is quite similar to subcl in set.mm. The condition on the existence
of the supremum is the same and the condition on the order is different
in the same way as other supremum theorems.

Also move the $d A z $. further up. Although this means it is present
for a few theorems which don't need it, generally it doesn't look like
it should be a problem (and many of the theorems in this scope do need
it).