Add gcdval to iset.mm
gcdval is stated as in set.mm but with a different proof which needs to
approach the existence of the supremum differently than in set.mm.

Includes gcdsupcl which is a lemma about the existence of the supremum.

Includes dvdsbnd which is somewhat similar to gcdcllem1 in terms of
expressing an upper bound for the divisors of an integer, but stated
differently (in a way which is convenient for the gcdsupcl proof).