add greatest common divisor to iset.mm (#2369)
* Add greatest common divisor to iset.mm

This is the gcd syntax and df-gcd , copied without change from set.mm.
Although we'll need some new supremum theorems to make df-gcd work,
I think it is probably as good a definition as any and as far as I can
tell the supremum part can be handled.

* Add zsupcl to iset.mm

This is a case where we can show a supremum exists.

Includes lemmas zsupcllemstep and zsupcllemex .

* Add gcdval to iset.mm

gcdval is stated as in set.mm but with a different proof which needs to
approach the existence of the supremum differently than in set.mm.

Includes gcdsupcl which is a lemma about the existence of the supremum.

Includes dvdsbnd which is somewhat similar to gcdcllem1 in terms of
expressing an upper bound for the divisors of an integer, but stated
differently (in a way which is convenient for the gcdsupcl proof).

* Add gcd0val to iset.mm

Copied without change from set.mm.

* Add gcdn0val to iset.mm

Copied without change from set.mm.

* Add gcdcllem1 , gcdcllem2 , gcdcllem3 to mmil.html

* Add gcdn0cl to iset.mm

The theorem is stated as in set.mm. The proof is in terms of gcdsupcl
however.

* Add zsssupcl to iset.mm

This is a case where we can prove the supremum of a set is a member of
that set.

* Add gcdsupex to iset.mm

This factors out part of the gcdsupcl proof for use elsewhere, and seems
necessary unless we have another way of expressing that the supremum
exists for use in theorems like supubti .

* Add gcdmndc to iset.mm

This factors out into a separate theorem a decidability condition which
we currently use once and which I expect we'll use a few more times
before we are done with gcd .

* Add gcddvds to iset.mm

The theorem is stated as in set.mm. The proof is different because of
the differences in how supremum theorems work.

* Add dvdslegcd to iset.mm

The theorem is stated as in set.mm. The proof is different because of
differences in how supremum works.

* Add nndvdslegcd to iset.mm

Copied without change from set.mm.

* Add gcdcl to iset.mm

The theorem is stated as in set.mm. The proof is basically the same as
the set.mm one except that excluded middle is replaced by gcdmndc .

* Add gcdnncl to iset.mm

Copied without change from set.mm

* Add gcdcld to iset.mm

Copied without change from set.mm.

* Add gcd2n0cl to iset.mm

Copied without change from set.mm

* Add zeqzmulgcd to iset.mm

Copied without change from set.mm

* Add divgcdz to iset.mm

Copied without change from set.mm.

* Add gcdf to iset.mm

Copied without change from set.mm.

* Add gcdcom to iset.mm

Copied without change from set.mm.

* Add divgcdnn to iset.mm

Copied without change from set.mm

* Add divgcdnnr to iset.mm

Copied without change from set.mm.

* Add gcdeq0 to iset.mm

The statement of the theorem is as in set.mm. The proof is very similar
to the set.mm one but we need to assert the deciability of one of the
propositions before using contraposition .

* Add gcdn0gt0 to iset.mm

The theorem is stated as in set.mm. The proof is mostly the same as the
set.mm one but in one part makes use of the fact that apartness and
negated equality are equivalent for integers.

* Add ddif to mmil.html

This belongs on the missing theorems list.

* Add gcd0id to iset.mm

The theorem is stated as in set.mm. The proof is similar to the set.mm
one but excluded middle needs to be changed to decidable equality for
integers.

* Add gcdid0 to iset.mm

Copied without change from set.mm.

* Add nn0gcdid0 to iset.mm

Copied without change from set.mm.

* Add gcdneg to iset.mm

The theorem is stated as in set.mm. The proof is almost the same as the
set.mm one but needs to change excluded middle to gcdmndc .

* Add neggcd to iset.mm

Copied without change from set.mm.

* Add pm4.57 to mmil.html

This is just oran commuted, which is already listed.

* Add orandc to iset.mm

This is oran from set.mm but for decidable propositions.

* Add gcdaddm to iset.mm

The proof is similar to the set.mm one with some changes:
- instead of having a lemma in inference form which is then turned into
deduction form using the deduction theorem, just do the whole thing in
deduction form
- replace excluded middle with decidability of a proposition which has
various connectives joining integer equalities

* Add gcdadd to iset.mm

Copied without change from set.mm.

* Add gcdid and ex-gcd to iset.mm

Copied without change from set.mm

* Shorten df-gcd comment in iset.mm

Remove the reference to dfgcd2 which is in the next section and which we
therefore do not have yet.

* Add gcd1 to iset.mm

The theorem is stated as in set.mm. The proof is the same as the set.mm
one except for the name of 1ne0 (instead of ax-1ne0 ).

* Add gcdabs to iset.mm

The theorem is stated as in set.mm. The proof is the set.mm one with
absor changed to qabsor (which is a very minor change since it is being
used on integers).

* Add gcdabs1 to iset.mm

The theorem is stated as in set.mm. The proof is the set.mm one with
absord changed to qabsord (which is a small change because we are
applying it to an integer).

* Add gcdabs2 to iset.mm

Copied without change from set.mm.

* Add modgcd and 1gcd to iset.mm

modgcd is stated as in set.mm and the proof is similar to the set.mm one
but with real changed to rational in several places.

1gcd is copied without changes.

* Add 6gcd4e2 to iset.mm

Copied without change from set.mm