Add eqinfti to iset.mm
This is similar to eqinf in set.mm. The statement of the theorem changes
the condition on the order from Or to tightness. The proof also is similar
to the set.mm one with the adjustments being for the condition on the
order (which, as in set.mm, does transfer from the original order to
its converse).