Significantly extend README.md
The original README.md said almost nothing.
Let's modify it so the README.md actually provides more detail.

I tried to capture the rule for approving merges per the list,
but with more precision. If you think I got it wrong, please
tell me how to fix it! I think it's important to document the
rule(s) for merging changes.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>