Move contribution text to CONTRIBUTING.md
Originally my contributing text was going to be a few sentences
so I was just going to put it in the README.
However, it's getting long, so per a recommendation from @jkingdon
I'm moving the contribution material into its own file.
I've added cross-links to make it easier to find.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>