infimum for iset.mm (#2374)
* Add infeq1 to iset.mm

Copied without change from set.mm.

* Add infeq1d to iset.mm

Copied without change from set.mm

* Add infeq1i , infeq2 , infeq3 to iset.mm

Copied without change from set.mm

* Add infeq123d to iset.mm

Copied without change from set.mm

* Add nfinf to iset.mm

Copied without change from set.mm

* Add infexd to mmil.html

* Add impel to iset.mm

Copied without change from set.mm

* Add eqinfti to iset.mm

This is similar to eqinf in set.mm. The statement of the theorem changes
the condition on the order from Or to tightness. The proof also is similar
to the set.mm one with the adjustments being for the condition on the
order (which, as in set.mm, does transfer from the original order to
its converse).

* Add eqinftid to iset.mm

This follows from eqinfti much as eqinfd follows from eqinf in set.mm,
with the proof also being analogous to the set.mm one.

* Add cnvti to iset.mm

This is a lemma which is helpful for proving eqinfti and most theorems
which want to prove an infimum theorem in terms of its supremum
counterpart.

* Add infvalti to iset.mm

Similar to infval in set.mm except with a different condition on the
order.

Includes lemma cnvinfex which lets us convert between supremem on the
converse, versus infimum on the original.

* Add infcllem to mmil.html

* Add infclti to iset.mm

This is infcl from set.mm with changes to the condition on the order.

* Add inflbti to iset.mm

This is inflb from set.mm with the usual changes for the conditions on
the order.

* Add infglbti to iset.mm

This is infglb from set.mm with the usual change to the condition on the
order.

* Add infglbb to mmil.html

* Add infnlbti to iset.mm

The theorem is stated as in set.mm except for the usual changes to the
conditions on the order. The proof is similar to the set.mm one except
that dfrex2 is changed to rexalim (because we are only using one
direction of the biconditional).

* Add infex to mmil.html

Same issues as infexd

* Add infminti to iset.mm

This is infmin from set.mm with the usual changes to the conditions on
the order.

* Add infmoti to iset.mm

The theorem is infmo from set.mm with the usual changes to the
conditions on the order. The proof is in terms of supmo which seemed
easier than trying to fix up a proof in terms of rmo4 .

* Add infeuti to iset.mm

This is infeu from set.mm with the usual change for the condition on the
order.

* Add frfi to mmil.html

* Add fimax2g and fimin2g to mmil.html

Includes fimaxg and fiming

* Add fiinfg , fiinf2g to mmil.html

* Add fiinfcl to mmil.html

* Add infltoreq to mmil.html

* Add infpr to mmil.html

* Add infsnti to iset.mm

This is infsn from set.mm with the usual change for the condition on
the order. The proof is in terms of supsnti .

* Add inf00 to iset.mm

Copied without change from set.mm

* Add infisoti to iset.mm

This is infiso from set.mm with the usual changes to the condition on
the order. The proof is in terms of supisoti , analogous to the set.mm
proof.

* Fix typo in inf00 comment

bas -> base

This is in iset.mm and set.mm