Significantly extend README.md (#2373)
* Significantly extend README.md

The original README.md said almost nothing.
Let's modify it so the README.md actually provides more detail.

I tried to capture the rule for approving merges per the list,
but with more precision. If you think I got it wrong, please
tell me how to fix it! I think it's important to document the
rule(s) for merging changes.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Emphasize early on how many verifiers we use

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Many small edits thanks to review by @bprzybocki

Benjamin (@bprzybocki) provided a lot of really nice editorial
fixes to the README update. This commit attempts to address them all.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Move contribution text to CONTRIBUTING.md

Originally my contributing text was going to be a few sentences
so I was just going to put it in the README.
However, it's getting long, so per a recommendation from @jkingdon
I'm moving the contribution material into its own file.
I've added cross-links to make it easier to find.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Remove unnecessary quote marks

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Update README/CONTRIBUTING per community feedback

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Fix link

A URL was followed by a period that meant "end of sentence",
which meant the generated link was wrong. Fix it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Tweak clarification per @tirix

@tirix wanted to make it clear that not-formally-verified
doesn't apply to Metamath proofs. I think this rewording does it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Tweak README for antecedents

The antecedent for "they" was too far away.
Repeat "set.mm and iset.mm" instead, to make it clear.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* CONTRIBUTING.md: Fix minor formatting problem

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* CONTIRBUTING.md: Tell reviewers to notice discouraged changes

It's not *wrong* to change a "discouraged" file, but it signals
something unusual is going on. Remind reviewers to look for this,
and check if it makes sense.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Move info on verifiers to VERIFIERS.md

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Fix pluralization

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Remove hyphen

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Fix typo

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>