Adjust CONTRIBUTING language about rewrapping
It is possible that most of the text introduced by "If you prefer
to run metamath yourself" can be removed or drastically pared
down, but at least for now maybe it is good to keep this for people who
are used to having it there? I don't think it is wrong per se.